build, ci, and repo workflows: modernize CMake setup and add PR/issue automation
Introduce structured GitHub issue templates for bug reports, suggestions, and RFCs
to make triage and larger design discussions more consistent. Add pull request
automation for blocked/stacked dependency chains so linked PRs can be labeled,
status-checked, and refreshed automatically when their dependencies are merged
or closed.

Modernize the top-level CMake configuration by moving away from broad global flag
string manipulation toward compile/link options and newer CMake facilities. Raise
the minimum required version to 3.25, switch MSVC debug information handling to
`CMAKE_MSVC_DEBUG_INFORMATION_FORMAT`, tighten release/security-oriented compiler
settings, and clean up platform-specific flag handling for MSVC, Clang, MinGW,
and Unix-like builds. Also update Qt deprecation defines, add explicit release
fortify flags for C, and keep the toml++ aarch64 workaround in place.

Refresh launcher metadata and platform integration details. Add a dedicated
launcher wiki URL, clarify dialog help URL behavior, fix Linux MIME info
installation to use the generated file, and add macOS asset catalog compilation
support through `actool` for newer Xcode-based icon packaging paths.

Simplify and substantially reorganize `CMakePresets.json` to reduce preset sprawl
and make the build surface easier to maintain. This aligns build entry points
more closely with the current packaging and CI model instead of carrying a very
large matrix of highly granular per-target presets.

Finally, trim release-note prose in the changelog by removing the GitLab migration
justification section so the changelog stays focused on shipped launcher changes
rather than platform strategy.

Overall, this commit improves repository governance, reduces build-system drift,
and makes cross-platform packaging/CI behavior more predictable.

Yes! I may genuinely love life, but this job has started to become fun and enjoyable
for me. Of course, that includes backporting, bug fixing, and adding features.

+++ launcher/ui/dialogs/skins/draw/BoxGeometry.cpp
@@ -306,11 +306,11 @@ namespace opengl
 
 		// Transfer vertex data to VBO 0
 		m_vertexBuf.bind();
-		m_vertexBuf.allocate(verticesData.constData(), verticesData.size() * sizeof(VertexData));
+		m_vertexBuf.allocate(verticesData.constData(), static_cast<int>(verticesData.size() * sizeof(VertexData)));
 
 		// Transfer index data to VBO 1
 		m_indexBuf.bind();
-		m_indexBuf.allocate(indices.constData(), indices.size() * sizeof(GLushort));
+		m_indexBuf.allocate(indices.constData(), static_cast<int>(indices.size() * sizeof(GLushort)));
 		m_indecesCount = indices.size();
 	}
 
@@ -325,11 +325,11 @@ namespace opengl
 
 		// Transfer vertex data to VBO 0
 		b->m_vertexBuf.bind();
-		b->m_vertexBuf.allocate(planeVertices.constData(), planeVertices.size() * sizeof(VertexData));
+		b->m_vertexBuf.allocate(planeVertices.constData(), static_cast<int>(planeVertices.size() * sizeof(VertexData)));
 
 		// Transfer index data to VBO 1
 		b->m_indexBuf.bind();
-		b->m_indexBuf.allocate(planeIndices.constData(), planeIndices.size() * sizeof(GLushort));
+		b->m_indexBuf.allocate(planeIndices.constData(), static_cast<int>(planeIndices.size() * sizeof(GLushort)));
 		b->m_indecesCount = planeIndices.size();
 
 		return b;
